moved regridding to packagebase for metaswap
